Abort program if either background task fails
This uses select!  instead of a try-join that would let background
threads fail and the program continue anyhow.

This was easiest to notice if the socket was already bound, as the
exporter would keep running even as it was unable to serve any data.

Now instead it properly fails and terminates cleanly, although maybe not
signalling failure exit code.