Rework error handling and logging
Sometimes tasks would get stuck and we could not see what was the cause,
as some would end with an Ok(())  as they didn't do anything wrong, but
in reality the channel they read data from terminated and caused the
issue.

This rewrites them to use an error, and adds error handling in the
tasks, removing the except() and unwrap() cases.

And we also re-introduce the timeout behaviour as that is needed.